pyproject.toml
[project]
name = "cmd-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "guedesf",email = "guedesf@google.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "ipykernel (>=7.2.0,<8.0.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-openai (>=1.1.12,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-core (>=1.2.21,<2.0.0)",
    "langchain-classic (>=1.0.3,<2.0.0)",
    "pyrogram (>=2.0.106,<3.0.0)",
    "tgcrypto (>=1.2.5,<2.0.0)",
    "supabase (>=2.28.3,<3.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "flask (>=3.1.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "cmd_agent", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
